backend/requirements.txt
Flask-Cors==3.0.10
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
filelock==3.9.0
Flask==2.2.3
imgaug==0.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
nibabel==3.2.1
numpy
opencv-python-headless==4.7.0.72
python-dotenv==0.21.1
Pillow==9.3.0
scipy==1.7.3
scikit-image==0.16.2
Shapely==2.0.1
SimpleITK==2.2.1
#sympy==1.11.1
webencodings==0.5.1
Werkzeug==2.2.3
wincertstore==0.2
tensorflow==2.5.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu
torchaudio==0.9.1
